+++ packages/aws-appsync/src/cache/offline-cache.ts
@@ -18,6 +18,8 @@ const WRITE_CACHE_ACTION = 'AAS_WRITE_CACHE';
 
 export interface OfflineCache extends NormalizedCache {
     rehydrated: boolean,
+    [NORMALIZED_CACHE_KEY]: any,
+    [METADATA_KEY]: { idsMap: object },
 }
 
 export default class MyCache extends InMemoryCache {
@@ -63,6 +65,12 @@ export default class MyCache extends InMemoryCache {
         this.store.dispatch(writeThunk(WRITE_CACHE_ACTION, {}));
 
         return super.reset();
+    }
+
+    getIdsMap() {
+        const { [METADATA_KEY]: { idsMap } } = this.store.getState();
+
+        return { ...idsMap };
     }
 }