+++ tutorials/react-offline-realtime-todos/src/GraphQLAllTodos.js
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag';
+
+export default gql`
+query {
+  listTodos {
+    items {
+      id
+      name
+      description
+      status
+      version
+    }
+  }
+}`
+