+++ tutorials/react-offline-realtime-todos/src/GraphQLDeleteTodo.js
@@ -1,0 +1,13 @@
+import gql from 'graphql-tag';
+
+export default gql`
+mutation($id: ID!) {
+  deleteTodo(input:{id: $id}){
+    id
+    name
+    description
+    status
+    version
+  }
+}`
+