+++ packages/aws-appsync-react/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.1.2"></a>
+## [1.1.2](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.1.1...aws-appsync-react@1.1.2) (2018-07-20)
+
+
+### Bug Fixes
+
+* **typescript:** Typescript improvements ([#191](https://github.com/awslabs/aws-mobile-appsync-sdk-js/issues/191)) ([712d089](https://github.com/awslabs/aws-mobile-appsync-sdk-js/commit/712d089))
+
+
+
+
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.1.0...aws-appsync-react@1.1.1) (2018-07-18)
+
+
+### Bug Fixes
+
+* **typescript:** Add types info ([#187](https://github.com/awslabs/aws-mobile-appsync-sdk-js/issues/187)) ([739470e](https://github.com/awslabs/aws-mobile-appsync-sdk-js/commit/739470e))
+
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.0.21...aws-appsync-react@1.1.0) (2018-07-17)
+
+
+### Features
+
+* **offline:** Offline helpers ([#184](https://github.com/awslabs/aws-mobile-appsync-sdk-js/issues/184)) ([0c0fe82](https://github.com/awslabs/aws-mobile-appsync-sdk-js/commit/0c0fe82))
+
+
+
+
 <a name="1.0.21"></a>
 ## [1.0.21](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.0.20...aws-appsync-react@1.0.21) (2018-07-10)